Create stringfy (passing objects as a parameter) in internal js.html
This HTML page creates a fruit object with properties name set to "Banana" and price set to 100. The JSON.stringify() method is used to convert the fruit object into a JSON-formatted string, which is then appended to the content of the paragraph with ID output. The result displayed is: The fruit object is {"name":"Banana","price":100}.