Create skipping elements of the nested array in internal js.html
This HTML page demonstrates array destructuring in JavaScript. The array arr contains the values [2, [3, 4], [9, 10]]. Using destructuring, it extracts the following values: a is assigned the first value 2, b is assigned the second value 4 (from the second sub-array [3, 4]), and c is assigned the value 10 (from the third sub-array [9, 10]). The result is displayed in the paragraph with ID output.