Create nested array destructing in internal js.html
This HTML page demonstrates array destructuring in JavaScript. The array arr contains the values [10, [15, 20], 30]. Using destructuring, it extracts the following values: a is assigned 10, b is assigned 15 (the first value in the second sub-array [15, 20]), c is assigned 20 (the second value in the second sub-array), and d is assigned 30. The result is displayed in the paragraph with ID output.