Create function and object (using the method) in internal js.html
This HTML page demonstrates the creation of user-defined objects in JavaScript using constructor functions. The Book function is defined as a constructor for creating book objects, which has properties title, author, and a method addPrice to set the price of the book. A new book object myBook is created with the title "Perl" and author "Mohtashim". The addPrice() method is called to set the price of the book to 100. The book's title, author, and price are then displayed in the div with the ID output.