Create function and object (using the constructor) in internal js.html
This HTML page demonstrates the use of a constructor function to create a Book object. The Book constructor takes title and author as parameters and assigns them to the object's properties. A new Book object myBook is created with the title "Perl" and author "Mohtashim". The title and author of the book are then displayed in the div with the ID output.