Create set objects or set method in internal js.html
This HTML page demonstrates the use of a Set in JavaScript. A Set is a collection of unique values, meaning it automatically removes duplicates. The code adds the values 60, 50, and 50 to the num_set set. Since sets do not allow duplicates, only 60 and 50 will be stored in the set. The for...of loop is used to iterate over the values in the set, and the values are displayed in the div with the ID output.