Create removing the set element in internal js.html
This HTML page demonstrates the use of the delete method in a JavaScript Set. The Set initially contains the values 100, 200, 300, 400, and 500. The delete method is used to remove the values 200 and 400 from the set. After the deletion, the remaining values in the set are iterated using the for...of loop and displayed in the div with the ID output.