Create union of two sets in internal js.html
This HTML page demonstrates the union of two sets, set1 and set2. The Set objects set1 and set2 contain car brands, and the union of these sets is created by using the spread operator (...) to combine the two sets into a new Set. The result is displayed in the div with the ID output.