Create difference of two sets in internal js.html
This HTML page calculates the difference between two sets, set1 and set2. It creates a copy of set1 in the diff set and then removes elements that are also present in set2. The result is displayed in the div with the ID output.

In this case, set1 contains ["BMW", "Audi", "TATA"], and set2 contains ["BMW", "TATA", "Honda", "Suzuki"]. After removing the elements present in set2 from set1, the difference will be:

Audi

This is because "BMW" and "TATA" are common to both sets, so only "Audi" remains in the difference.