Create weak set objects 1 in internal js.html
In this code, a WeakSet is created containing two objects (obj1 and obj2). The WeakSet allows the inclusion of objects, but it does not prevent them from being garbage collected if there are no other references to them.

The code checks if the WeakSet contains obj1 and obj2 using the has() method. It then outputs whether each object is present in the WeakSet.